@@ -145,7 +145,7 @@
     "favicons-webpack-plugin": "^0.0.7",
     "file-loader": "0.9.0",
     "html-loader": "0.4.4",
-    "html-webpack-plugin": "2.24.1",
+    "html-webpack-plugin": "^2.28.0",
     "husky": "^0.13.2",
     "image-webpack-loader": "2.0.0",
     "imports-loader": "0.6.5",
@@ -167,7 +167,7 @@
     "sinon": "^2.1.0",
     "style-loader": "0.13.1",
     "url-loader": "0.5.7",
-    "webpack": "2.2.0-rc.3",
+    "webpack": "^2.3.0",
     "webpack-appversion-plugin": "^1.0.6",
     "webpack-dev-middleware": "1.9.0",
     "webpack-hot-middleware": "2.15.0",
