@@ -35,7 +35,7 @@ export default function configureStore(initialState = {}, history) {
     };
 
     Array.prototype.push.apply(enhancers, [
-      require('../utils/devtools.component').instrument(),
+      require('../utils/devtools.component').default.instrument(),
       persistState(getDebugSessionKey(), (state) => fromJS(state)),
     ]);
   }
