@@ -21,7 +21,7 @@ const plugins = [
     to: 'fixtures',
   }]),
   new webpack.HotModuleReplacementPlugin(),
-  new webpack.NoErrorsPlugin(),
+  new webpack.NoEmitOnErrorsPlugin(),
   new HtmlWebpackPlugin({
     // Inject all files that are generated by webpack, e.g. bundle.js
     inject: true,
